+++ examples/react-dom/.eslintrc.js
@@ -1,6 +1,4 @@
 /* @flow */
 /* eslint-env commonjs */
 
-module.exports = {
-    'extends': ['../../react-dom.js'],
-}
+module.exports = {'extends': ['../../react-dom.js']}